# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pulse API
  version: 0.1.0
  description: |
    API REST de Pulse — herramienta personal de gestión de contenido y publicaciones
    en redes sociales, multi-workspace.

    ## Autenticación
    - **Sesión web**: cookie JWT (no implementado todavía en v0.1.0).
    - **Integraciones (SESIÓN 2)**: header `Authorization: Bearer sk_ws_xxx...` con scopes.

    ## Aislamiento de workspace
    Todas las rutas bajo `/api/v1/w/:slug/*` están scopeadas al workspace cuyo slug
    aparece en la URL. El middleware `workspaceScope` valida que el workspace existe
    (404 si no) y que el caller tiene acceso (403 si no).
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.pulse.local
    description: Production (placeholder)
paths:
...
  /api/v1/w/{slug}/campaigns:
    get:
      tags: [Settings]
      parameters: [{ $ref: "#/components/parameters/Slug" }]
      responses: { "200": { description: OK } }
    post:
      tags: [Settings]
      parameters: [{ $ref: "#/components/parameters/Slug" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                objective: { type: string }
                startAt: { type: string, format: date-time }
                endAt: { type: string, format: date-time }
                kpiName: { type: string }
                kpiTarget: { type: number }
      responses: { "201": { description: Created } }
...
  /api/v1/w/{slug}/api-keys:
    get:
      tags: [Settings]
      parameters: [{ $ref: "#/components/parameters/Slug" }]
      responses: { "200": { description: OK } }
    post:
      tags: [Settings]
      summary: Crear API key
      description: |
        Devuelve el token completo en la respuesta. Es la única vez que se ve;
        guárdalo de forma segura.
      parameters: [{ $ref: "#/components/parameters/Slug" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string }
                scopes:
                  type: array
                  items: { type: string, enum: [ingest, read_brain, read_metrics] }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  token: { type: string, example: "sk_ws_a8f3c9d1_xxxxxxxxxxxxxxxxxxxxxxxx" }
                  keyPrefix: { type: string }
                  scopes: { type: array, items: { type: string } }
                  warning: { type: string }
...
components:
  parameters:
    Slug:
      in: path
      name: slug
      required: true
      schema: { type: string, pattern: "^[a-z0-9-]+$" }
...
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: Bearer token `sk_ws_xxx`. Las API keys se crean desde Settings > API Keys.
...